driver/py/requirements.txt
aiohttp>=3.9.0
python-socketio>=5.11.0
pynput>=1.7.6
pyperclip>=1.8.2
qrcode>=7.4.2
netifaces>=0.11.0